+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `Span`, for splitting an `Iterable` into contiguous elements matching a predicate
 - `MagnetizeBy` and `Magnetize`, for grouping elements by pairwise predicate tests
 - `Both`, for dually applying two functions and producing a `Tuple2` of their results
+- `Lens#both`, for dually focusing with two lenses at once
 
 ### Deprecated
 - `Either#trying` in favor of `Try#trying`